Added new qrutils submodule for useful widgets. Some widgets from graphicsUtils migrated there
