+++ plugins/robots/interpreters/trikKitInterpreter/src/trikDisplayWidget.ui
@@ -304,7 +304,7 @@ QPushButton:pressed{
     </spacer>
    </item>
    <item row="1" column="2" rowspan="5" colspan="2">
-    <widget class="graphicsUtils::PaintWidget" name="display" native="true">
+    <widget class="qReal::ui::PaintWidget" name="display" native="true">
      <property name="sizePolicy">
       <sizepolicy hsizetype="Preferred" vsizetype="Preferred">
        <horstretch>13</horstretch>
@@ -317,9 +317,9 @@ QPushButton:pressed{
  </widget>
  <customwidgets>
   <customwidget>
-   <class>graphicsUtils::PaintWidget</class>
+   <class>qReal::ui::PaintWidget</class>
    <extends>QWidget</extends>
-   <header>qrutils/graphicsUtils/paintWidget.h</header>
+   <header>qrutils/widgets/paintWidget.h</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
