+++ qrutils/qrutils.qrc
@@ -5,5 +5,6 @@
         <file>graphicsWatcher/icons/zoomOut_btn.png</file>
         <file>graphicsWatcher/icons/zoomIn_btn.png</file>
         <file>graphicsWatcher/icons/reset_btn.png</file>
+        <file>widgets/icons/find.svg</file>
     </qresource>
 </RCC>
